Rename the scope_manager to variable_manager and move to new module
