+++ src/CMakeLists.txt
@@ -16,6 +16,9 @@ add_executable(
     functions.cpp
     bytecode.cpp
     runtime.cpp
+
+    compilation/type_manager.cpp
+    compilation/variable_manager.cpp
 )
 
 target_include_directories(anzu PRIVATE .)
