+++ src/compilation/type_manager.cpp
@@ -1,0 +1,5 @@
+#include "type_manager.hpp"
+
+namespace anzu {
+    
+}
