+++ src/compilation/type_manager.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+namespace anzu {
+    
+}
