--- src/utility/stack.hpp
@@ -1,13 +1,0 @@
-#pragma once
-#include "utility/common.hpp"
-#include "utility/memory.hpp"
-
-#include <utility>
-#include <array>
-#include <vector>
-#include <cstdint>
-
-namespace anzu {
-
-
-}
